Fixed windows build - deactivated saving for now

+++ modules/filmstro_ffmpeg/filmstro_ffmpeg_FFmpegVideoWriter.h
@@ -123,7 +123,7 @@ private:
     int                     subtitleStreamIdx;
 
     int                     sampleRate;
-    int                     channelLayout;
+    int64_t                 channelLayout;
 
     int                     videoWidth;
     int                     videoHeight;
